move the config engine
