+++ build.gradle
@@ -62,6 +62,7 @@ sourceSets {
         java {
             srcDir 'api'
             srcDir 'util/src/org/commcare/util/screen'
+            srcDir 'util/src/org/commcare/util/engine'
             compileClasspath += main.output
         }
     }
