+++ util/src/org/commcare/util/cli/ApplicationHost.java
@@ -7,7 +7,7 @@ import org.commcare.data.xml.DataModelPullParser;
 import org.commcare.suite.model.FormIdDatum;
 import org.commcare.suite.model.SessionDatum;
 import org.commcare.suite.model.StackFrameStep;
-import org.commcare.util.CommCareConfigEngine;
+import org.commcare.util.engine.CommCareConfigEngine;
 import org.commcare.util.CommCarePlatform;
 import org.commcare.session.SessionFrame;
 import org.commcare.util.mocks.CLISessionWrapper;
