+++ util/src/org/commcare/util/cli/CliValidateCommand.java
@@ -1,7 +1,7 @@
 package org.commcare.util.cli;
 
 import org.apache.commons.cli.ParseException;
-import org.commcare.util.CommCareConfigEngine;
+import org.commcare.util.engine.CommCareConfigEngine;
 import org.javarosa.core.util.externalizable.LivePrototypeFactory;
 import org.javarosa.core.util.externalizable.PrototypeFactory;
